Crate solana_short_vec

Compact serde-encoding of vectors with small length.

Structs

ShortU16

Same as u16, but serialized with 1 to 3 bytes. If the value is above 0x7f, the top bit is set and the remaining value is stored in the next bytes. Each byte follows the same pattern until the 3rd byte. The 3rd byte may only have the 2 least-significant bits set, otherwise the encoded value will overflow the u16.

ShortVec

Functions

decode_shortu16_len

Return the decoded value and how many bytes it consumed.

deserialize

If you don’t want to use the ShortVec newtype, you can do ShortVec deserialization on an ordinary vector with the following field annotation:

serialize

If you don’t want to use the ShortVec newtype, you can do ShortVec serialization on an ordinary vector with the following field annotation: